fixed 8a3aae7 from https://gitee.com/lixing0101/hiviewdfx_hidumper/pulls/326 update GetCpuUsage interface

+++ interfaces/native/innerkits/include/dump_broker_cpu_proxy.h
@@ -29,7 +29,7 @@ public:
 public:
     // Used for dump request
     int32_t Request(DumpCpuData &dumpCpuData) override;
-    int32_t GetCpuUsageByPid(int32_t pid, int &cpuUsage) override;
+    int32_t GetCpuUsageByPid(int32_t pid, double &cpuUsage) override;
 private:
     static inline BrokerDelegator<DumpBrokerCpuProxy> delegator_;
 };
